refactor: Extract GroupPanelSummary into own file
